MDL-69374 core_course: Handle guest access in can_view_course_info()
When disabling the 'moodle/category:viewcourselist' capability for the guest role
the guest user couldn't access any course where the guest enrolment method was enabled
(i.e. guest access is allowed). The above mentioned capability overruled guest access settings
for a course.
The fix adds additional checking if the guest enrolment is enabled for a course
in the 'can_view_course_info()' method of the 'core_course_category' class
in /course/classes/category.php. And if enabled the guest user and any auth'ed user
may access the course.