MDL-64648 enrol: fixed setting of 'notifyall'
When creating a course a manual enrolment instance is added by default.
The instance settings should inherit the settings of manual enrolment plugin.
The 'expirynotify' setting wasn't inherited correctly in case it had
'Enroller + Enrolled' value.
Also, the 'notifyall' setting wasn't set correctly (as it's value is based on
the 'expirynotify' value) when adding an enrolment instance and when editing it as well
(the latest is mentioned in MDL-61811 also).
The fix addresses both problems and comprises several steps to modify the code:
1. The visibility of the 'get_expirynotify_options()' method of
the manual enrolment plugin was set to public to enable it's calling
in plugin's settings.php to populate options for the 'expirynotify' setting.
2. The 'get_expirynotify_options()' is called in the settings.php
to populate options for 'expirynotify'. That brings more consistency to the code.
3. Setting of 'notifyall' was taken from the 'add_default_instance()' method
and put to the 'add_instance()' method. That's due to 'add_instance()' is finally
called by 'add_default_instance()' anyway and 'add_instance()' is also called
when an enrolment instance is added through the interface, so 'notifyall' should be
set there as well.
4. The 'update_instance()' method was modified to set 'notifyall' as well.
Finally, the fix appeared to fix MDL-61811 as well.
A unit test was added to /enrol/manual/tests/lib_test.php to test setting
of 'notifyall'.